appengine/requirements.txt
-i https://pypi.python.org/simple
cachecontrol==0.12.5
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.0
firebase-admin==3.2.0
flask==1.1.1
google-api-core[grpc]==1.14.3 ; platform_python_implementation != 'PyPy'
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-firestore==1.6.0 ; platform_python_implementation != 'PyPy'
google-cloud-storage==1.23.0
google-cloud-tasks==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos[grpc]==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.25.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.3
markupsafe==1.1.1
msgpack==0.6.2
protobuf==3.11.0
pyasn1-modules==0.2.7
pyasn1==0.4.8
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
uritemplate==3.0.0
urllib3==1.25.7
werkzeug==0.16.0
